update fundamental organization of repo
key changes:
- data files are organized in 4-level hierarchy as follows:
  = hazard type (currently: earthquake, hurricane, flood)
  = asset type (currently: building, transportation_network, water_network)
  = scale (component, subassembly, portfolio)
  = method

- Data files corresponding to a specific method are stored in a folder
at each ending branch of the above hierarchy. The following files are
envisioned:
  = fragility
  = consequence
  = loss

- Each of the above options has two corresponding files. The model
parameters are stored in a CSV while the metadata is stored in a JSON.

- The raw data that was used to generate these files and any other data
source deemed important can be stored in a `data_sources` folder within
each method's own folder. Currently, we've put the raw data under an
`input files` folder within `data_sources` and, when available, a
generator script is also placed there to support re-generating the
data files for the method, should the raw data get updated.

- Additional resources envisioned to be included in the methods folder:
  = mapping scripts that map models (i.e., components, archetypes) to
  asset characteristics
  = combination tables that support the combination of losses from
  multiple types of demands (e.g., hurricane wind and surge)