Fix a few bugs and issues with the Hazus flood loss models
- Use 'loss' instead of 'consequence' in the filename to signal that
these are indeed loss functions
- Fix a bug in the generator script that yielded overwritten models in
the csv file. Now we get a larger number and comprehensive list of
models including those for structural and inventory losses
- Modify the ID for models to start with the type (structural,
inventory, or contents), followed by the function ID that is referenced
in the Hazus technical manual. This should facilitate mapping buildings
to archetypes.
- Remove 'structure' and 'inventory' from the description part of the
ID. I kept 'equipment' and 'equipment/inventory' as that seemed to add
useful information.
- Sort the entries in the CSV by their index to facilitate manual lookup