Update doc generator script to exclude Hazus Hurricane and Hazus EQ subassembly for the time being.
Hurricane is excluded because the generated docs would be too large for
Github to host. We are working on a solution.

Hazus EQ subassembly is excluded because the library is identical to the
component library. We'll need to review this workflow and decide how to
handle having a unique mapping script without changes to the core
library.