Fix a bug in the ID of Hazus hurricane components
Masonry Low-Rise Malls had their joist spacing attribute as a float
instead of an int which led to an extra zero in the ID. This issue
affected both the original and coupled datasets and both the fragility
and the consequence/loss functions.

All affected files are replaced with this commit.

The commit also includes the bugfix in the generator script.