Add a list of core model files
This list is used by Pelicun to identify and pull only the core model files that are needed for the calculation.
The JSON files are only used in the PBE app to display component information. They will be removed from this list as soon as the PBE app migrates to another approach and does not need them to be part of Pelicun.